Only run test for specific server versions

+++ gocd/pipeline_action_test.go
@@ -10,50 +10,47 @@ import (
 func testPipelineServiceUnPause(t *testing.T) {
 	for n, test := range []struct {
 		name          string
-		v             *ServerVersion
+		v             *ServerVersionRange
 		confirmHeader string
 		acceptHeader  string
 	}{
 		{
 			name:          "server-version-14.3.0",
-			v:             &ServerVersion{Version: "14.3.0"},
+			v:             newServerVersionRangeFromString("1.0.0", "14.3.0"),
 			confirmHeader: "Confirm",
 			acceptHeader:  apiV0,
 		},
 		{
 			name:          "server-version-18.3.0",
-			v:             &ServerVersion{Version: "18.3.0"},
+			v:             newServerVersionRangeFromString("14.3.0", "18.3.0"),
 			confirmHeader: "X-GoCD-Confirm",
 			acceptHeader:  apiV1,
 		},
 	} {
 		t.Run(test.name, func(t *testing.T) {
 			if runIntegrationTest(t) {
+				if runOnlyForServerVersionRange(t, test.v) {
 
-				ctx := context.Background()
-				pipelineName := fmt.Sprintf("test-pipeline-un-pause%d", n)
+					ctx := context.Background()
+					pipelineName := fmt.Sprintf("test-pipeline-un-pause%d", n)
 
-				err := test.v.parseVersion()
-				assert.NoError(t, err)
+					pausePipeline, _, err := intClient.PipelineConfigs.Create(ctx, mockTestingGroup, &Pipeline{
+						Name: pipelineName,
+						Materials: []Material{{
+							Type: "git",
+						}},
+						Stages: buildMockPipelineStages(),
+					})
+					assert.NoError(t, err)
+					assert.Equal(t, nil, pausePipeline)
 
-				cachedServerVersion = test.v
+					pp, _, err := intClient.Pipelines.Pause(ctx, pipelineName)
+					assert.NoError(t, err)
+					assert.True(t, pp)
 
-				pausePipeline, _, err := intClient.PipelineConfigs.Create(ctx, mockTestingGroup, &Pipeline{
-					Name: pipelineName,
-					Materials: []Material{{
-						Type: "git",
-					}},
-					Stages: buildMockPipelineStages(),
-				})
-				assert.NoError(t, err)
-				assert.Equal(t, nil, pausePipeline)
-
-				pp, _, err := intClient.Pipelines.Pause(ctx, pipelineName)
-				assert.NoError(t, err)
-				assert.True(t, pp)
-
-				deleteResponse, _, err := intClient.PipelineConfigs.Delete(ctx, pipelineName)
-				assert.Equal(t, "", deleteResponse)
+					deleteResponse, _, err := intClient.PipelineConfigs.Delete(ctx, pipelineName)
+					assert.Equal(t, "", deleteResponse)
+				}
 			}
 		})
 	}
